Provide a unified resource cleanup facility.
The CL type now has a Writer component that accumulates actions to
release resources. The idea is that after you runCL a CL action, you get
a result and the composite cleanup action. This frees you from having to
build up your own cleanup action. This facility is also supported by the
Image and Buffer types that come with allocation and initialization
functions that automatically register for cleanup. The upshot is that if
you use Images or Buffers persistently, you get the cleanup action as a
single entity that is built for you automatically.