Experimenting with memory mapped async operations.
I'm still not certain how I want this to work. It's very tempting to
kick off the asynchronous operation in a separate thread, but I
shouldn't lose sight of the fact that the main point here is to operate
on mapped memory rather than demanding the driver write to my pointer.